Add werkzeug ProxyFix X-Forwarded-Prefix support
The base url is not the same as the blueprint mount point because the blueprint url is relative to the application context root.
So either the jinja template needs to be altered to use relative urls, or the base_url should be determined per request.
This commit changes the base_url to url_for(".show"), this makes the UI work for both proxied and direct requests.